Fixed first run with wrong settings, improved exit flow.

+++ src/sovok_pvr_client.cpp
@@ -42,16 +42,23 @@ SovokPVRClient::SovokPVRClient(CHelper_libXBMC_addon *addonHelper, CHelper_libXB
     m_sovokTV(addonHelper, sovokLogin, sovokPassword),
     m_inputBuffer(NULL),
     m_isTimeshiftEnabled(false),
-    m_shouldAddFavoritesGroup(true),
-    m_CacheDir(s_DefaultCacheDir)
+    m_shouldAddFavoritesGroup(true)
 {
-    if(!m_addonHelper->RemoveDirectory(m_CacheDir.c_str()))
-        m_addonHelper->Log(LOG_ERROR, "Failed to remove cache folder");
+    SetTimeshiftPath(s_DefaultCacheDir);
 }
 
 SovokPVRClient::~SovokPVRClient()
 {
     CloseLiveStream();
+}
+
+void SovokPVRClient::SetTimeshiftPath(const std::string& path){
+    const char* nonEmptyPAth = (path.empty()) ? s_DefaultCacheDir : path.c_str();
+    if(m_addonHelper->DirectoryExists(path.c_str())) {
+        m_CacheDir = path;
+        if(!m_addonHelper->RemoveDirectory(m_CacheDir.c_str()))
+            m_addonHelper->Log(LOG_ERROR, "Failed to remove cache folder");
+            }
 }
 
 PVR_ERROR SovokPVRClient::GetEPGForChannel(ADDON_HANDLE handle, const PVR_CHANNEL& channel, time_t iStart, time_t iEnd)
